# allama/tables/internal_router.py
"""Internal router for table operations (SDK/UDF use)."""

from __future__ import annotations

from datetime import datetime
from typing import Any, Literal
from uuid import UUID

import orjson
from asyncpg import DuplicateTableError
from fastapi import APIRouter, HTTPException, Query, status
from pydantic import BaseModel, Field
from pydantic_core import to_jsonable_python
from sqlalchemy.exc import DBAPIError, ProgrammingError

from allama import config
from allama.auth.dependencies import ExecutorWorkspaceRole
from allama.db.dependencies import AsyncDBSession
from allama.exceptions import AllamaNotFoundError
from allama.expressions.functions import tabulate
from allama.logger import logger
from allama.tables.common import coerce_optional_to_utc_datetime
from allama.tables.enums import SqlType
from allama.tables.schemas import (
    TableColumnCreate,
    TableColumnRead,
    TableCreate,
    TableRead,
    TableRowInsert,
    TableRowInsertBatch,
)
from allama.tables.service import TablesService

router = APIRouter(
    prefix="/internal/tables", tags=["internal-tables"], include_in_schema=False
)


class TableCreateRequest(BaseModel):
    name: str
    columns: list[TableColumnCreate] = Field(default_factory=list)
    raise_on_duplicate: bool = True


class TableLookupRequest(BaseModel):
    columns: list[str]
    values: list[Any]
    limit: int | None = None


class TableExistsRequest(BaseModel):
    columns: list[str]
    values: list[Any]


class TableSearchRequest(BaseModel):
    search_term: str | None = None
    start_time: datetime | None = None
    end_time: datetime | None = None
    updated_before: datetime | None = None
    updated_after: datetime | None = None
    offset: int = Field(default=0, ge=0)
    limit: int = Field(default=100, ge=1)


class TableRowUpdate(BaseModel):
    data: dict[str, Any]


TableDownloadFormat = Literal["json", "ndjson", "csv", "markdown"]


@router.get("")
async def list_tables(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
) -> list[dict[str, Any]]:
    """List all tables for the workspace."""
    service = TablesService(session, role=role)
    tables = await service.list_tables()
    return [table.to_dict() for table in tables]


@router.post("", status_code=status.HTTP_201_CREATED)
async def create_table(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    params: TableCreateRequest,
) -> dict[str, Any]:
    """Create a new table."""
    service = TablesService(session, role=role)
    try:
        table = await service.create_table(
            TableCreate(name=params.name, columns=params.columns)
        )
    except ProgrammingError as exc:
        # Drill down to the root cause
        while (cause := exc.__cause__) is not None:
            exc = cause
        if isinstance(exc, DuplicateTableError):
            if params.raise_on_duplicate:
                raise HTTPException(
                    status_code=status.HTTP_409_CONFLICT,
                    detail="Table already exists",
                ) from exc
            await session.rollback()
            table = await service.get_table_by_name(params.name)
        else:
            logger.error("Unexpected error creating table", error=str(exc))
            raise HTTPException(
                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                detail="An error occurred while creating the table",
            ) from exc
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc
    return table.to_dict()


@router.get("/{table_name}/metadata", response_model=TableRead)
async def get_table_metadata(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
) -> TableRead:
    """Get table metadata by name."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    index_columns = await service.get_index(table)
    return TableRead(
        id=table.id,
        name=table.name,
        columns=[
            TableColumnRead(
                id=column.id,
                name=column.name,
                type=SqlType(column.type),
                nullable=column.nullable,
                default=column.default,
                is_index=column.name in index_columns,
                options=column.options,
            )
            for column in table.columns
        ],
    )


@router.post("/{table_name}/lookup")
async def lookup_rows(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    params: TableLookupRequest,
) -> list[dict[str, Any]]:
    """Lookup rows matching column/value pairs."""
    if (
        params.limit is not None
        and params.limit > config.ALLAMA__MAX_ROWS_CLIENT_POSTGRES
    ):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=(
                f"Limit cannot be greater than {config.ALLAMA__MAX_ROWS_CLIENT_POSTGRES}"
            ),
        )
    service = TablesService(session, role=role)
    try:
        return await service.lookup_rows(
            table_name,
            columns=params.columns,
            values=params.values,
            limit=params.limit,
        )
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc


@router.post("/{table_name}/exists")
async def exists_rows(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    params: TableExistsRequest,
) -> bool:
    """Check if any rows match column/value pairs."""
    service = TablesService(session, role=role)
    try:
        return await service.exists_rows(
            table_name,
            columns=params.columns,
            values=params.values,
        )
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc


@router.post("/{table_name}/search")
async def search_rows(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    params: TableSearchRequest,
) -> list[dict[str, Any]]:
    """Search rows in a table with optional filters."""
    if params.limit > config.ALLAMA__MAX_ROWS_CLIENT_POSTGRES:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=(
                f"Limit cannot be greater than {config.ALLAMA__MAX_ROWS_CLIENT_POSTGRES}"
            ),
        )
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    try:
        return await service.search_rows(
            table,
            search_term=params.search_term,
            start_time=coerce_optional_to_utc_datetime(params.start_time),
            end_time=coerce_optional_to_utc_datetime(params.end_time),
            updated_before=coerce_optional_to_utc_datetime(params.updated_before),
            updated_after=coerce_optional_to_utc_datetime(params.updated_after),
            limit=params.limit,
            offset=params.offset,
        )
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc


@router.post("/{table_name}/rows", status_code=status.HTTP_201_CREATED)
async def insert_row(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    params: TableRowInsert,
) -> dict[str, Any]:
    """Insert a row into a table."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    try:
        return await service.insert_row(table, params)
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc


@router.post("/{table_name}/rows/batch", status_code=status.HTTP_201_CREATED)
async def insert_rows_batch(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    params: TableRowInsertBatch,
) -> int:
    """Insert multiple rows into a table atomically."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    try:
        return await service.batch_insert_rows(table, params.rows, upsert=params.upsert)
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc
    except DBAPIError as exc:
        detail = str(exc.__cause__ or exc)
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=f"Database error: {detail}",
        ) from exc


@router.patch("/{table_name}/rows/{row_id}")
async def update_row(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    row_id: UUID,
    params: TableRowUpdate,
) -> dict[str, Any]:
    """Update a row in a table."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    try:
        return await service.update_row(table, row_id=row_id, data=params.data)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc
    except ValueError as exc:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(exc),
        ) from exc


@router.delete("/{table_name}/rows/{row_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_row(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    row_id: UUID,
) -> None:
    """Delete a row from a table."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    try:
        await service.delete_row(table, row_id=row_id)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc


@router.get("/{table_name}/download")
async def download_table(
    *,
    role: ExecutorWorkspaceRole,
    session: AsyncDBSession,
    table_name: str,
    format: TableDownloadFormat | None = Query(default=None),
    limit: int = Query(default=1000, ge=1, le=1000),
) -> list[dict[str, Any]] | str:
    """Download table data as JSON, NDJSON, CSV, or Markdown."""
    service = TablesService(session, role=role)
    try:
        table = await service.get_table_by_name(table_name)
    except AllamaNotFoundError as exc:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=str(exc),
        ) from exc

    rows = await service.list_rows(table=table, limit=limit)
    json_safe_rows = to_jsonable_python(rows, fallback=str)

    if format is None:
        return json_safe_rows
    if format == "json":
        return orjson.dumps(json_safe_rows).decode()
    if format == "ndjson":
        return "\n".join([orjson.dumps(row).decode() for row in json_safe_rows])
    if format in {"csv", "markdown"}:
        return tabulate(json_safe_rows, format)
    return tabulate(json_safe_rows, format)