# docs/openapi.yml
openapi: 3.1.0
info:
  title: Allama API
  summary: Allama API
  description: Allama is the open source automation platform for enterprise. You can
    operate Allama in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Allama Founders
    email: founders@allama.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /tables:
    get:
      tags:
      - tables
      summary: List Tables
      description: List all tables.
      operationId: tables-list_tables
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableReadMinimal'
                title: Response Tables-List Tables
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tables
      summary: Create Table
      description: Create a new table.
      operationId: tables-create_table
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tables/{table_id}:
    get:
      tags:
      - tables
      summary: Get Table
      description: Get a table by ID.
      operationId: tables-get_table
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - tables
      summary: Update Table
      description: Update table metadata.
      operationId: tables-update_table
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tables
      summary: Delete Table
      description: Delete a table.
      operationId: tables-delete_table
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /tables/{table_id}/rows:
    get:
      tags:
      - tables
      summary: List Rows
      description: Get a row by ID.
      operationId: tables-list_rows
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          description: Maximum number of rows to return
          default: 100
          title: Limit
        description: Maximum number of rows to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          description: Number of rows to skip
          default: 0
          title: Offset
        description: Number of rows to skip
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableRowRead'
                title: Response Tables-List Rows
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tables
      summary: Insert Row
      description: Create a new row in a table.
      operationId: tables-insert_row
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableRowInsert'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tables/{table_id}/rows/{row_id}:
    get:
      tags:
      - tables
      summary: Get Row
      description: Get a row by ID.
      operationId: tables-get_row
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: row_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Row Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tables
      summary: Delete Row
      description: Delete a row from a table.
      operationId: tables-delete_row
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: row_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Row Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tables/{table_id}/rows/batch:
    post:
      tags:
      - tables
      summary: Batch Insert Rows
      description: 'Insert multiple rows into a table atomically.


        All rows will be inserted in a single transaction. If any row fails,

        the entire batch will be rolled back.'
      operationId: tables-batch_insert_rows
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: table_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Table Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableRowInsertBatch'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRowInsertBatchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SqlType:
      type: string
      enum:
      - TEXT
      - INTEGER
      - NUMERIC
      - BOOLEAN
      - TIMESTAMP
      - TIMESTAMPTZ
      - JSONB
      - UUID
      title: SqlType
      description: Supported SQL types.
    TableColumnCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: The name of the column
        type:
          $ref: '#/components/schemas/SqlType'
          maxLength: 100
          minLength: 1
          description: The SQL type of the column
        nullable:
          type: boolean
          title: Nullable
          default: true
        default:
          anyOf:
          - {}
          - type: 'null'
          title: Default
      type: object
      required:
      - name
      - type
      title: TableColumnCreate
      description: Create model for a table column.
    TableColumnRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          title: Name
        type:
          $ref: '#/components/schemas/SqlType'
        nullable:
          type: boolean
          title: Nullable
          default: true
        default:
          anyOf:
          - {}
          - type: 'null'
          title: Default
        is_index:
          type: boolean
          title: Is Index
          default: false
      type: object
      required:
      - id
      - name
      - type
      title: TableColumnRead
      description: Definition for a table column.
...
    TableCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: The name of the table
        columns:
          items:
            $ref: '#/components/schemas/TableColumnCreate'
          type: array
          title: Columns
          description: The columns of the table
      type: object
      required:
      - name
      title: TableCreate
      description: Create model for a table.
    TableRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          title: Name
        columns:
          items:
            $ref: '#/components/schemas/TableColumnRead'
          type: array
          title: Columns
      type: object
      required:
      - id
      - name
      - columns
      title: TableRead
      description: Read model for a table.
    TableReadMinimal:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: TableReadMinimal
      description: Read model for a table.
    TableRowInsert:
      properties:
        data:
          type: object
          title: Data
        upsert:
          type: boolean
          title: Upsert
          default: false
        natural_keys:
          items:
            type: string
          type: array
          title: Natural Keys
          description: The columns of the table to use for upsert
      type: object
      required:
      - data
      title: TableRowInsert
      description: Insert model for a table row.
    TableRowInsertBatch:
      properties:
        rows:
          items:
            type: object
          type: array
          title: Rows
      type: object
      required:
      - rows
      title: TableRowInsertBatch
      description: Request body for batch inserting rows.
    TableRowInsertBatchResponse:
      properties:
        rows_inserted:
          type: integer
          title: Rows Inserted
      type: object
      required:
      - rows_inserted
      title: TableRowInsertBatchResponse
      description: Response for batch insert operation.
    TableRowRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: true
      type: object
      required:
      - id
      - created_at
      - updated_at
      title: TableRowRead
      description: Read model for a table row.
    TableUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
          description: The name of the table
      type: object
      title: TableUpdate
      description: Update model for a table.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-allama-service-key
...